+++ program-libs/interface/tests/parser_props.rs
@@ -14,7 +14,6 @@ use zolana_event::MessageData;
 use zolana_interface::instruction::instruction_data::{
     deposit::{
         DepositEntry, DepositIxData, EncryptedRingDepositData, RingDepositEntry, RingDepositIxData,
-        UtxoData,
     },
     merge_ring::{MergeRingIxData, MergeRingIxDataRef},
     merge_transact::{
@@ -381,20 +380,17 @@ proptest! {
         owner in any::<[u8; 32]>(),
         envelope_seed in any::<[u8; 32]>(),
         amount in any::<u64>(),
-        utxo_data in prop::option::of((any::<[u8; 32]>(), prop::collection::vec(any::<u8>(), 0..300))),
         memo in prop::option::of(prop::collection::vec(any::<u8>(), 0..300)),
         ring_data_hash in any::<[u8; 32]>(),
         ring_data in prop::collection::vec(any::<u8>(), 0..300),
         cut in any::<prop::sample::Index>(),
         trailing in any::<u8>(),
     ) {
-        let utxo_data = utxo_data.map(|(data_hash, data)| UtxoData { data_hash, data });
         let entry = DepositEntry {
             asset_index: 0,
             view_tag,
             owner,
             amount,
-            utxo_data: utxo_data.clone(),
             memo: memo.clone(),
         };
         let deposit = DepositIxData {
@@ -408,7 +404,6 @@ proptest! {
                 view_tag,
                 owner_utxo_hash: owner,
                 amount,
-                data_hash: utxo_data.map(|data| data.data_hash),
                 ring_data_hash,
                 encrypted: EncryptedRingDepositData {
                     tx_viewing_pk: [envelope_seed[0]; 33],